Experiment: Create a separate tests/fake/test_gsp_fake.py test case module
